Extract extension methods class
